added testing infrastructure with automated tests

+++ run_spinner.sh
@@ -16,17 +16,32 @@ print_target_pid() {
     echo "Target PID: $TARGET_PID"
 }
 
-# Main loop for accepting commands
-while true; do
+print_options() {
     echo ""
     echo "Enter a command:"
+    echo "0) Show options"
     echo "1) Kill target pid"
     echo "2) Send a SIGUSR1 signal to it"
     echo "3) Print target pid"
-    echo "4) Exit"
+    echo "4) forku - create doppelganger"
+    echo "5) Exit"
+}
+
+print_options
+
+# Main loop for accepting commands
+while true; do
+    if ! kill -0 $TARGET_PID 2>/dev/null; then
+        echo "[*] Spinner process got killed..."
+        exit 1
+    fi
+
     read -p "> " cmd
 
     case $cmd in
+        0)
+            print_options
+            ;;
         1)
             kill $TARGET_PID
             echo "Killed target process with PID $TARGET_PID"
@@ -40,12 +55,18 @@ while true; do
             print_target_pid
             ;;
         4)
+            echo "taskset -c 0 $FORKU_UTIL $TARGET_PID"
+            taskset -c 0 $FORKU_UTIL $TARGET_PID
+            ;;
+        5)
             break
             ;;
         *)
             echo "Invalid command"
             ;;
     esac
+
+    sleep 0.2
 done
 
 echo "Exiting script."
